requirements.txt
pandas
beautifulsoup4
requests
pyyaml